Move SubChannel.SetHandler tests to subchannel_test.go
